model/requirements.txt
ultralytics>=8.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
PyYAML>=6.0
